Use int instead of numpy.int64

+++ mpmorph/analysis/clustering_analysis.py
@@ -112,7 +112,7 @@ class ClusteringAnalyzer(object):
         return avg_dist
 
     def find_clusters(self):
-        _sites = set(np.arange(len(self.cluster_neighbors)))
+        _sites = set(np.arange(len(self.cluster_neighbors), dtype=int))
         clusters = []
         while _sites:
             site = _sites.pop()
